shouldi: Run bandit and safety if Python
This patch changes shouldi to take a git repo instead of a Python
package name. It clones the repo, then determines what programming
lanaguage it's written in. If it's written in Python (currenly decides
this by checking if there is a setup.py at the root of the repo), then
it kicks of the pervious dataflow it used to use to analyize Python
packages.

Run with:

$ cd examples/shouldi
$ python setup.py test -s tests.test_cli.TestCLI.test_install

Signed-off-by: John Andersen <johnandersenpdx@gmail.com>